requirements.txt
openant>=0.4.0
pyusb>=1.2.1
ha-mqtt-discoverable>=0.18.0
paho-mqtt>=2.1.0 